Add files via upload
JDK 11 and Java FX 11 are dependencies for this application. 
Maybe later I can update this by reconstructing the projects in JDK 10 or 8 (both are bundled with JavaFX) so that I can create and upload executable .jar files. Having JavaFX as an external library seems to make IntelliJ IDEA confused when creating artifacts for the .jar file.